Reduce padding on body and title text.
Due to the intrinsic padding in TextViews, the UX padding values between
the body and title text, above the title text and below the body text
causes the line item to be too tall.

Instead, reduce this padding so that it achieves the required results.

Bug: 87828353
Test: Test on sample app with TextListItem. Compare a single line item
with a double line item that has one line of body text and ensure they
are the same height.

Change-Id: I9e3d469481b2651e53ca36a45e8570aa2a0a943b