Attempt to fix flaking StaggeredGridLayoutManagerTest
Two different flakes fixed:

1. forceLayoutOnDetach

I think the test was flaking because an assertion was
happening outside of the thread that would cause
it to be true. I've simply moved the assertion to happen
right after the call that should cause it to be true to
avoid any possible race conditions.

2. innerGapHandling

Here, I couldn't determine what was causing the flake
other than layout time possibly requiring more than 2
seconds.  I've just bumped that to 10. I also separated
out two tests that were occuring in one run to avoid the
possibility of the first interfering with the second.

Bug: 73660981
Bug: 73260264
Fixes: 73660981
Fixes: 73260264
Test: StaggeredGridLayoutManagerTest

Change-Id: I3054afcbe9a4ca92ae4f8d0e1c2807f1ce037cd5