Prevent database lock on Schedulers.schedule.
In a previous changelist (ag/4502245), I fixed a race
condition where work was scheduled and tried to run
before it was committed to the database.  For this
reason, work needs to be fully committed before the
Scheduler#schedule call happens.  There is still one
case where this is happening - EnqueueRunnable may
end up calling CancelWorkRunnable, and the scheduling
may happen inside EnqueueRunnable's transaction.

This can actually lead to a deadlock - somebody may
have locked the database and is calling synchronized
Schedulers#schedule, and another thread may be trying
to grab the database from inside Schedulers#schedule.

Removing the synchronized on Schedulers#schedule is
also something we need to do, but I want to check
this in as a standalone CL since it fixes the bug
by itself.  The synchronized change will be a
secondary CL.

Bug: 111801342
Test: Ran through Muzei example in bug.
Change-Id: I720ba78ae3e5b7e487427f9c6c1a556b27fe00b4