Attempt to fix flaking StaggeredGridLayoutManagerTest.
I believe the issue was that the time that was allotted
to scroll the distance desired was simply not long
enough when running on devices under unknown strain.

I did two things to decrease the flakiness.
1: Decreased the distance being scrolled.
2: Made the size of the RV fixed.  (The item heights
were based on the RV's height, the RV's height was
based on the screen's, and the scroll velocity is
fixed, so the larger the screen, the longer the
scroll would take. Now that it is fixed, the
time it takes to scroll should also be fixed).

Bug: 112512487
Test: androidx.recyclerview.widget.StaggeredGridLayoutManagerTest
	#temporaryGapHandling

Change-Id: I4607dc207fc3bbd32096f08ef7b3508df8c1981a