Explicitly label WM components as direct boot unaware.
Because WM defaults to a database in encrypted storage, it
cannot run while in direct boot.  By explicitly labelling the
components this way, they cannot inherit a directBootAware
flag from the merged <application> xml.

In the future, we will provide a version of WorkManager that
is direct-boot aware.  See the tracking feature request at
https://issuetracker.google.com/112773820  This is something
that might land post-1.0.

Bug: 112665532
Test: Ran some integration tests in a test app.
Change-Id: Ide87b8e4c5dfefb88a54c3ce4c45262335f47e96