[R4A] Allow views to define children lambdas
This is needed for some views (like a List/Recycler for example) where
we want to have an attribute that would be a "composable" function, and
it makes sense for that attribute to take the "children" slot so that
we can use KTX body lambdas to define those attributes. In these cases,
the normal addView/removeView for those views would make no sense anyway.

Change-Id: I6e5ea1a2aef9e5a63f55ab2dd8196ef9ed343e21