Fix flaking RecyclerViewLayoutTest.
Test seems to be flaking on Nexus 4 with API 15.
After some investigation, it turns out there is some
real weird stuff happening with some RV tests related
to, test SimpleItemAnimators,
BaseRecyclerViewInstrumentationTest, and API 15, where
on API 15, RV doesn't post item change updates to
"OnAnimation" and instead requests a layout, and
a grid lock occurs between waiting for animations to
finish and waitForIdleSync().

This fix doesn't address that larger question, but does
prevent the gridlock for a couple of tests.

Fixes: 113048055
Fixes: 113290948
Test: androidx.recyclerview.widget.RecyclerViewLayoutTest
      #smoothScroll_whenSmoothScrollerStops_destinationReached

Change-Id: I38282220af59606165a14659a8d38c7692da75d1