Change the fill color of media notification button drawables
In API 19, MediaSessionService2's default media notification had
black buttons on black background, which makes the buttons less visible.

It is okay for lollipop and higher APIs, since the buttons' color is
automatically changed. However, it doesn't change in API < 21.

This CL changes the button drawables' color from black to white
so that the buttons become more visible.

Bug: 112448525
Test: Checked the Notification on API 19/21/24/28
Change-Id: I270448b89ce1bc2eb2fee63146012a39e3dc033f