Add ListenableFuture dependency in WorkManager.
This is an extremely naive implementation and WILL need to be
changed.  I just want to get the gradle file updated and a
rudimentary example that passes tests.

This CL specifically returns a ListenableFuture from
NonBlockingWorker#onStartWork and observes it to chain the
result to WorkerWrapper#onWorkFinished.  It also exposes
the background Executor for androidx.work.TaskExecutor.

And to reiterate, this is a naive, non-final implementation
and will not be the one shipped in a release.

Test: Ran all tests.
Change-Id: Ia97ec192544f643915206d6ce3e691f7e695936e