Add a first integration test
It uses a CraneView to draw an image on the screen. We look at the
canvas where the image is drawn, and ensure that it has (approximately)
the correct pixels. Due to the bitmap being scaled (and filtered?)
before being drawn to the canvas, I could not use the jetpack image to
compare, as it had lots of noise in the pixels being drawn. Instead, I
used a new image, having four quadrants of different colours, and I am
checking that the thing being drawn preserves the 25% pixel frequency.

Note that I do not own any copyright for the picture; I will manufacture
one myself if people think this test looks fine.

Test: ran the test
Change-Id: I1a3d5bcd8de407f2d03bfbbcf80f8e963a9784e0