Add annotation processing features to dejetifier.
** This is just merge from beta **

We have generated annotation.zip files in our artifacts to be used with
Lint. This allows dejetifier to correctly process these files. The case
turned out to be a bit more tricky as for instance media has
annotations.xml files in both old and new packages so merging was
neccessary to suport this case.

Also deletes the old source transformer config file as it was confusing
which one is the correct one.

Test: Added tests for rewirting and merging
Bug: b/68490704

Change-Id: I9f33d357c39073810b6ee80bde6faed6c677e904
Merged-in: I3b3d4d5da056852ae5108dcdc518eaa9c07df8bf