Remove methods for setting text appearance.
Instead, clients should rely on a ViewBinder for individual ListItem
changes. If they wish to change the appearance for multiple, then it
should be modified through the theme.

Also, remove the setters in the ListItemAdapter because these values can
just be read off the theme with attribute queries in the XML themselves.

Bug: 116816039
Test: Verify text appearance can be set on SupportCarDemos with
ViewBinders. Otherwise, they should take the values specified in the
theme.
Change-Id: Id54c6cc6c8590fa6060f2ae6415b74e20e7b0dcb