Make SettableFuture unhidden as ResolvableFuture.
After discussion with the Guava team, there is a need
to expose a lightweight ListenableFuture implementation
for cases when the developer has to give us back a
ListenableFuture.  We don't want to clash with Guava's
naming, so we are changing the name to
ResolvableFuture.

This change affects media2, but all changed files are
hidden so this will not affect any public APIs or
change any contracts.

Test: Ran media and concurrent's tests.
Change-Id: I676bfb38ece303617e820a39f2d403cbdfffa9e7