Improve Invalidation Tracker Memory Table
This change updates the in-memory table used by the invalidation
tracker to be of table_id & invalidated status, instead of version &
table_id. This improves the tracker by reducing the amount of write
operations performed by the trigger on large transactions. Instead the
trigger just flips the invalidated status to ON reducing the write
operations made by the trigger which in turns also reduces the rollback
log file.

Added a room-benchmark artifact containing a benchmark test that can
be used to demonstrate the performance improvement of this change.

Test: ./gradlew room:room-runtime:test \
                room:integration-tests:testapp:cC

Change-Id: Ie9254a050d25f8d7dcdf7427f71018960a310d38