Remove legacy exclude in init.gradle
Historically, we had a list of projects that we did not want to
configure. We moved most of our configuration to plugins, but we
left the excludes. The exclude list got smaller over time and now
is just one project. noto-emoji-compat does not even have tests.
Remove the early return.

Test: ./gradlew buildOnServer
Change-Id: I22c05ebd7afbf8f716113528575d0bab2fc47574