Expose AppBarConfiguration APIs in navigation-ktx
Previously, navigation-ktx methods totally hid the
AppBarConfiguration object from the API surface.

While convenient, this leads to issues where the API
surface continues to change whenever AppBarConfiguration
changes. It also prevented developers from reusing the
same AppBarConfiguration object in other APIs, such
as the NavController.navigateUp(AppBarConfiguration) API
needed for ActionBar integration.

By providing a AppBarConfiguration extension constructors
to encapsulate the AppBarConfiguration.Builder, we retain
much of the improved API surface while making it clear that
we're using AppBarConfiguration as the primitive for
all of these APIs.

The Kotlin extensions now mirror the NavigationUI API surface,
with a method that takes a DrawerLayout (without a default
value to prevent ambiguous calls) and a separate method that
takes an AppBarConfiguration.

Test: testapp still works
BUG: 117333663
Change-Id: I46c58731913a514a373783b6e791420272e2c83b