117529548: MediaMetadataCompat.Builder does not support custom keys
This CL adds a new attribute, "open", to the typedef annotation,
which lets you indicate that a given typedef is "open": other
values are allowed, not just the explicitly enumerated constants.

Normally, typedefs are considered closed so if an API is invoked
passing any other constant, lint and the IDE will flag this as
an invalid value.

The MediaMetadataCompat library was using the @StringDef annotation
on an API which should also allow other values, which meant lint
was flagging these other values as invalid. With this new mechanism,
the API is annotated as open, and lint has been updated to look for
this additional attribute and if so silently allow other values.

(You might then wonder why bother annotating the API at all. There
is still value in this; for example, the IDE drives code completion
suggestions from these annotations, so Ctrl-Space on one of these
APIs will suggest all the included APIs. Similarly, metalava
automatically generates documentation from the APIs, etc.)

Change-Id: I47446cb27f18a5d5bf5e8abb69a60cc875f4bdfc
Fixes: 117529548
Test: Tests added in Android Lint