MediaSessionCompat: Remove reflection to MediaSession.mCallback
A contructor of MediaSessionCompat which gets a framework MediaSession
as an argument uses reflection to check whether the given session
already has a callback or not. When the framework session has
no callback, then a dummy MediaSessionCompat.Callback is set.

This CL removes this logic, and mentions this new behavior in the
Javadoc of MediaSessionCompat#fromMediaSession().

Bug: 118670049
Test: ./media/version-compat-tests/runtest.sh 1
Change-Id: I221eb559b7d16d233e682436ec90cc73767937e7