Add ReplacementDependency table, migration, and process startup code.
When we REPLACE work, it may continue executing even though new work
with that name will also execute.  We want to wait until it is finished
executing (since some Workers don't honor onStopped).  Because we
delete the database entries of REPLACEd work, this table will keep
track of any running workers and act as a source of truth before the
newly-enqueued work can run.

If the process dies, we clear this table on startup since those old
workers can no longer be running by definition.

The next CL will populate and use the table.

Bug: 117640278
Test: Added tests for migration and cleanup.
Change-Id: If172c7610df383f3fe63de4d91326290f111355f