Fix MediaPlayer2StateTest to run before P
Increase the timeout on initial preparation. When running a set of these
tests, the first time the resource is loaded preparation takes longer to
complete. For now, just increase the timeout (a later change should
ideally fix whatever is causing initially loading the asset to be slow,
and reduce the timeout back to one second).

Call selectTrack with a valid (non-video) track index. MediaPlayer2's
documentation states that "Currently, only timed text tracks or audio
tracks can be selected via this method.". The ExoPlayer-based player
implementation throws an IllegalArgumentException when trying to select
track 0 which is a video track.

Fix a field name typo.

Bug: 80232248
Test: Ran tests locally in Android Studio
Change-Id: I1bf82402d4a5eddb57ec1db0e1bce67b0871d29b