Remove NonNull annotation for TransitionValues.view
API council fix request. We can't really enforce TransitionValues.view to be @NonNull, so let's remove it for now. I am not sure about marking it as Nullable instead as then I will be forced to add null checks into a lot of usages, but it is not expected behavior. It was originally designed to threat this field as a NonNull, and it will be created only by our framework, 3-party developers should never create it manually (except for a testing purpose). So I would better treat this field being null as a runtime exception(NPE) while we are keeping this deprecated constructor.

Bug: 119620689
Test: n/a
Change-Id: I4412a56f31be811cadee1b61bb5a59a34bf1b23f