Remove NavDestination.navigate() and ref to Navigator
Provide a more clear separation of responsibilities
between a NavDestination and the Navigator itself by
removing the navigate() method from NavDestination.

This also allows us to remove the reference to the
creating Navigator from NavDestination, making it
safe to save NavGraph objects across configuration
changes and make it possible to dynamically change
the Navigator associated with the NavDestination
at runtime.

Test: New NavigatorProviderTest, existing tests pass
BUG: 111278637
Change-Id: Ic0d2ce02f1094a5d342dc62ad4f59f5c6e083926