[wv-supportlib] Support a :dev feature suffix.
:dev suffixed features will only be available if the device is
debuggable. This makes it possible to develop support library changes in
multiple steps, and then expose the feature with a final CL that removes
the suffix, without running the risk that applications will
inadvertently release code depending on the behaviour of an in-progress
implementation of a feature.

This is paired with chromium CL: https://crrev.com/c/1349327

Bug: 120148383
Test: ./gradlew :webkit:connectedAndroidTest
Change-Id: I0839e5e16364a3901d8376f69b67e59006fe8ee6