Enable lintMinDepVersions in buildOnServer
Runs lintMinDepVersionsDebug as part of buildOnServer. Aside from
enforcing the classic lint rules, lint will be more strict with
the RestrictedApi check given that a bug prevented it from
detecting those when depending on a local project version of a
library, but as we move more towards depending on versioned
artifacts, the check will start detecting restricted api
violations.

Current violations have been suppressed, and should be addressed
in the future.

Test: ./gradlew bOS and ./gradlew lintMinDepVersionsDebug
Change-Id: I18cf2799ac84f28db5b5b19baeec04c4c26d0732