ViewPager2 should now present the same actions and events to an accessibility service as ViewPager1.
Essentially, have the recyclerView treated like ViewPager1.
Use the same className, and make sure it populates the relevant event info so announcements can be made correctly.

Removed RecyclerView classname, because it is not being used by any service, it shouldn't be used by any service, it doesn't do what is expected because of progaurd, and it was preventing this fix from working, .

Test: Ran with ViewPager2 demo app and TalkBack
Change-Id: I78f89eddd8f94fdb651d0e89b4ee3f7bb382ee88