Signal EoS using -1 return in DataSourceCallback
MediaPlayerTest.testPlaybackFailsIfMedia2DataSourceReturnsAnError is
flaky because it may time out if ExoPlayer reads the entire source
before the test instructs its data source callback to return a negative
value from readAt.

The documentation of DataSourceCallback says that a -1 return value
signals both EoS and an error. Since errors are already signaled via
throwing, this change removes the ambiguity and uses -1 to indicate
reaching the end of the stream. See also aosp/832162 for context.

Once errors are signaled via throwing the flaky test is no longer needed
and can be removed, like we did for MediaPlayer2.

Bug: 118402459
Bug: 120622976
Bug: 80232248
Test: Ran tests in Android Studio
Change-Id: I3743703dd64963db63aa8339f6e648423a69234f